MIgrate to explicit rotation on-demand
Move away from the idea of relying on a background goroutine to handle
rotation, where every client was relying on the background goroutine
to not fail. This method lacked a clear way to communicate any
error directly to the client/caller.

The new method is to do the key rotation "on-demand", when a key
is actually requested and we detect that the current one is expired.